MyApiSolution/Services/OrderService.cs
namespace MyApiSolution.Services;

public interface IOrderService
{
    int PlaceOrder(int customerId, List<int> productIds);
    bool CancelOrder(int orderId);
    List<int> GetOrderDetails(int orderId);
}

public class OrderService
{
    public int PlaceOrder(int customerId, List<int> productIds)
    {
        // Stub: Add logic here
        return 0; // Placeholder
    }

    public bool CancelOrder(int orderId)
    {
        // Stub: Add logic here
        return false; // Placeholder
    }

    public List<int> GetOrderDetails(int orderId)
    {
        // Stub: Add logic here
        return new List<int>(); // Placeholder
    }

}